Lifted search term states up to App component
